fix more warnings in libYAML

+++ Sources/CYaml/src/api.c
@@ -1205,7 +1205,7 @@ yaml_document_add_scalar(yaml_document_t *document,
     SCALAR_NODE_INIT(node, tag_copy, value_copy, length, style, mark, mark);
     if (!PUSH(&context, document->nodes, node)) goto error;
 
-    return document->nodes.top - (int)document->nodes.start;
+    return (int)document->nodes.top - (int)document->nodes.start;
 
 error:
     yaml_free(tag_copy);
@@ -1250,7 +1250,7 @@ yaml_document_add_sequence(yaml_document_t *document,
             style, mark, mark);
     if (!PUSH(&context, document->nodes, node)) goto error;
 
-    return document->nodes.top - (int)document->nodes.start;
+    return (int)document->nodes.top - (int)document->nodes.start;
 
 error:
     STACK_DEL(&context, items);
@@ -1295,7 +1295,7 @@ yaml_document_add_mapping(yaml_document_t *document,
             style, mark, mark);
     if (!PUSH(&context, document->nodes, node)) goto error;
 
-    return document->nodes.top - (int)document->nodes.start;
+    return (int)document->nodes.top - (int)document->nodes.start;
 
 error:
     STACK_DEL(&context, pairs);
